requirements.txt
# LLM
groq>=0.9.0

# Embeddings
langchain-huggingface>=0.0.3
sentence-transformers>=2.6.1

# Vector store
chromadb>=0.5.0

# Document loading
pypdf>=4.0.0
requests>=2.32.0
beautifulsoup4>=4.12.0

# API
fastapi>=0.115.0
uvicorn>=0.30.0

# UI
streamlit>=1.37.0

# Utils
pandas>=2.0.0
python-dotenv>=1.0.1
pydantic>=2.0.0

# Tests
pytest>=8.0.0
